+++ packages/stacked/example/lib/app/app.router.dart
@@ -6,10 +6,12 @@
 
 // ignore_for_file: public_member_api_docs
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
 
 import '../ui/bottom_nav/bottom_nav_example.dart';
+import '../ui/details/details_view.dart';
 import '../ui/home/home_view.dart';
 import '../ui/stream_view/stream_counter_view.dart';
 
@@ -17,10 +19,12 @@ class Routes {
   static const String homeView = '/';
   static const String bottomNavExample = '/bottom-nav-example';
   static const String streamCounterView = '/stream-counter-view';
+  static const String detailsView = '/details-view';
   static const all = <String>{
     homeView,
     bottomNavExample,
     streamCounterView,
+    detailsView,
   };
 }
 
@@ -31,6 +35,7 @@ class StackedRouter extends RouterBase {
     RouteDef(Routes.homeView, page: HomeView),
     RouteDef(Routes.bottomNavExample, page: BottomNavExample),
     RouteDef(Routes.streamCounterView, page: StreamCounterView),
+    RouteDef(Routes.detailsView, page: DetailsView),
   ];
   @override
   Map<Type, StackedRouteFactory> get pagesMap => _pagesMap;
@@ -53,5 +58,28 @@ class StackedRouter extends RouterBase {
         settings: data,
       );
     },
+    DetailsView: (data) {
+      var args = data.getArgs<DetailsViewArguments>(
+        orElse: () => DetailsViewArguments(),
+      );
+      return CupertinoPageRoute<dynamic>(
+        builder: (context) => DetailsView(
+          key: args.key,
+          name: args.name,
+        ),
+        settings: data,
+      );
+    },
   };
+}
+
+/// ************************************************************************
+/// Arguments holder classes
+/// *************************************************************************
+
+/// DetailsView arguments holder class
+class DetailsViewArguments {
+  final Key key;
+  final String name;
+  DetailsViewArguments({this.key, this.name});
 }