+++ packages/stacked/example/lib/main.dart
@@ -1,8 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'main.router.dart';
+import 'package:new_architecture/app/app.locator.dart';
+import 'package:stacked_services/stacked_services.dart';
 
-void main() {
+import 'app/app.router.dart';
+
+Future main() async {
+  await setupLocator();
   runApp(MyApp());
 }
 
@@ -11,7 +15,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      navigatorKey: Get.key,
+      navigatorKey: StackedService.navigatorKey,
       onGenerateRoute: StackedRouter().onGenerateRoute,
       theme: ThemeData(
         primarySwatch: Colors.blue,