+++ packages/stacked/example/lib/services/shared_preferences_service.dart
@@ -1,8 +1,11 @@
 class SharedPreferencesService {
   static SharedPreferencesService _instance;
+  SharedPreferences _sharedPreferences;
 
-  static Future<SharedPreferencesService> getInstance() {
+  static Future<SharedPreferencesService> getInstance() async {
     if (_instance == null) {
+      // Initialise the asynchronous shared preferences
+      _sharedPreferences = await SharedPreferences.getSharedPrefs();
       _instance = SharedPreferencesService();
     }