+++ packages/stacked_generator/lib/src/generators/enums/dependency_type.dart
@@ -1,4 +1,4 @@
-enum ServiceType {
+enum DependencyType {
   Factory,
   LazySingleton,
   Singleton,