+++ packages/stacked_generator/lib/src/generators/getit/services_config.dart
@@ -1,9 +1,9 @@
-import 'package:stacked_generator/src/generators/getit/service_config.dart';
+import 'package:stacked_generator/src/generators/getit/dependency_config.dart';
 
 /// Describes the configuration of a service to be registered with get_it
 class ServicesConfig {
   /// A list of services to be registered with the get_it locator
-  final List<ServiceConfig> services;
+  final List<DependencyConfig> services;
 
   ServicesConfig({this.services});
 }