+++ packages/stacked_generator/lib/src/generators/getit/stacked_getit_generator.dart
@@ -6,8 +6,8 @@ import 'package:source_gen/source_gen.dart';
 import 'package:stacked/stacked_annotations.dart';
 import 'package:stacked_generator/import_resolver.dart';
 import 'package:stacked_generator/src/core/logger.dart';
-import 'package:stacked_generator/src/generators/enums/serivce_type.dart';
-import 'package:stacked_generator/src/generators/getit/service_config.dart';
+import 'package:stacked_generator/src/generators/enums/dependency_type.dart';
+import 'package:stacked_generator/src/generators/getit/dependency_config.dart';
 import 'package:stacked_generator/src/generators/getit/services_config.dart';
 import 'package:stacked_generator/utils.dart';
 
@@ -27,13 +27,13 @@ class StackedGetItGenerator extends GeneratorForAnnotation<StackedApp> {
     var libs = await buildStep.resolver.libraries.toList();
     var importResolver = ImportResolver(libs, element.source.uri.path);
 
-    final servicesConfig = stackedApplication.read('services').listValue;
+    final servicesConfig = stackedApplication.read('dependencies').listValue;
 
-    List<ServiceConfig> services = List<ServiceConfig>();
+    List<DependencyConfig> services = List<DependencyConfig>();
     // Convert the services config into Services configuration
     for (final serviceConfig in servicesConfig) {
-      final serialisedServiceConfig = _readServiceConfig(
-        serviceConfig: serviceConfig,
+      final serialisedServiceConfig = _readDependencyConfig(
+        dependencyConfig: serviceConfig,
         importResolver: importResolver,
       );
       services.add(serialisedServiceConfig);
@@ -42,45 +42,45 @@ class StackedGetItGenerator extends GeneratorForAnnotation<StackedApp> {
     return GetItLocatorGenerator(ServicesConfig(services: services)).generate();
   }
 
-  ServiceConfig _readServiceConfig({
-    DartObject serviceConfig,
+  DependencyConfig _readDependencyConfig({
+    DartObject dependencyConfig,
     ImportResolver importResolver,
   }) {
-    var serviceReader = ConstantReader(serviceConfig);
+    var dependencyReader = ConstantReader(dependencyConfig);
     // Get the type of the service that we want to register
-    final serviceClassType = serviceReader.read('service').typeValue;
+    final dependencyClassType = dependencyReader.read('classType').typeValue;
 
     throwIf(
-      serviceClassType == null,
-      '🛑 No service class Type defined for ${serviceConfig.toString()}. Please make sure that any of the services provided to the services list in the StackedApp annotation has a service provided. Please see the documentation for stacked_generator if you don\'t know what this means.',
+      dependencyClassType == null,
+      '🛑 No depedency class Type defined for ${dependencyConfig.toString()}. Please make sure that any of the services provided to the services list in the StackedApp annotation has a service provided. Please see the documentation for stacked_generator if you don\'t know what this means.',
     );
 
-    final classElement = serviceClassType.element as ClassElement;
+    final classElement = dependencyClassType.element as ClassElement;
 
     throwIf(
       classElement == null,
-      '🛑 ${serviceClassType.getDisplayString()} is not a class element. All services should be classes. We don\'t register individual values for global access through the locator. Make sure the value provided as your service type is a class.',
+      '🛑 ${dependencyClassType.getDisplayString()} is not a class element. All services should be classes. We don\'t register individual values for global access through the locator. Make sure the value provided as your service type is a class.',
     );
 
     // Get the import of the class type that's defined for the service
     final import = importResolver.resolve(classElement);
 
-    final className = serviceClassType.getDisplayString();
+    final className = dependencyClassType.getDisplayString();
 
     // NOTE: This can be used for actual dependency inject. We do service location instead.
     final constructor = classElement.unnamedConstructor;
 
-    final serviceType = _getServiceType(serviceReader);
+    final serviceType = _getDependencyType(dependencyReader);
     String presolveFunction;
-    if (serviceType == ServiceType.PresolvedSingleton) {
-      final presolveUsing = serviceReader.read('presolveUsing');
+    if (serviceType == DependencyType.PresolvedSingleton) {
+      final presolveUsing = dependencyReader.read('presolveUsing');
       final presolveObject = presolveUsing.objectValue.toFunctionValue();
       presolveFunction = presolveObject.displayName;
       log.d(
           'presolveUsing:$presolveUsing objectValue:${presolveObject.toString()} presolveFunction:$presolveFunction');
     }
 
-    return ServiceConfig(
+    return DependencyConfig(
       className: className,
       import: import,
       type: serviceType,
@@ -88,20 +88,20 @@ class StackedGetItGenerator extends GeneratorForAnnotation<StackedApp> {
     );
   }
 
-  ServiceType _getServiceType(ConstantReader serviceReader) {
+  DependencyType _getDependencyType(ConstantReader serviceReader) {
     if (serviceReader.instanceOf(TypeChecker.fromRuntime(Factory))) {
-      return ServiceType.Factory;
+      return DependencyType.Factory;
     }
     if (serviceReader.instanceOf(TypeChecker.fromRuntime(Singleton))) {
-      return ServiceType.Singleton;
+      return DependencyType.Singleton;
     }
     if (serviceReader.instanceOf(TypeChecker.fromRuntime(LazySingleton))) {
-      return ServiceType.LazySingleton;
+      return DependencyType.LazySingleton;
     }
     if (serviceReader.instanceOf(TypeChecker.fromRuntime(Presolve))) {
-      return ServiceType.PresolvedSingleton;
+      return DependencyType.PresolvedSingleton;
     }
 
-    return ServiceType.Singleton;
+    return DependencyType.Singleton;
   }
 }